Only make symlink if it is installed by npm

+++ tools/setup.sh
@@ -7,5 +7,7 @@ node tools/init-admin.js
 mkdir uploads
 mongoimport -d $DB -c role --drop < tools/initial-data/role
 cd ../..
-rm -f simaya
-ln -s node_modules/simaya simaya
+if [ -d node_modules/simaya ];then
+    rm -f simaya
+    ln -s node_modules/simaya simaya
+fi
